Remove the double-hit at the start of a pattern when using JACK transport.
In pattern mode, there's several redundant relocates to frame 0 in the
transport.  The result is that when using the JACK transport, we get
two, consecutive, process() cycles with pos->frame == 0 and
JackTransportRolling.  This means we play the first period twice.
This affected pattern mode only, and happened whether H2 was the
transport master or slave.  The double-hit also only happened at the
_start_ of the pattern.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@443 3b3fb362-3133-0410-aa15-cf69e0a59cb7