Trigger note release instead of abrupt cutoff on note-off events.
(Thanks to Krzysztof Foltman for the patch.)

Issue was reported to the forums [1] that bass drum kicks on TR808
created a clicking sound when used with a MIDI controller.  In the
sampler, NOTE OFF events create an abrupt mute on the instrument,
which creates a high-frequency click.  This patch instead triggers a
release on the instrument, respecting the mute group of the
instrument.

[1] http://www.hydrogen-music.org/forum/?action=show_thread&fid=0&thread=1034&page=1

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@889 3b3fb362-3133-0410-aa15-cf69e0a59cb7