Raise max A, D, and R setting from 10,000 frames to 100,000 frames.
Thanks to Krzysztof Foltman for the patch.

Before, the relationship between the Attack, Decay and Release
parameters and the rotary knobs that set them was:

    rotary = __param / 10000.0
    __param = rotary * 10000.0

This patch changes this to:

    rotary = sqrt( __param / 100000.0 )
    __param = rotary * rotary * 100000.0

So, not only is the max ADR parameter now 100,000 samples (typ. 2.26
sec @ 44.1 kHz), but the knob now works on a quadratic scale.  Values
close to zero will hold roughly the same values that they had before
the patch.  Here's a table for comparison:

        BEFORE             AFTER
     =============     =============
    rotary  __param | rotary  __param
    ------  ------- | ------  -------
     0.0         0  |  0.0         0
     0.05      500  |  0.05      250
     0.1      1000  |  0.1      1000
     0.2      2000  |  0.2      4000
     0.3      3000  |  0.3      9000
     0.4      4000  |  0.4     16000
     0.5      5000  |  0.5     25000
     0.6      6000  |  0.6     36000
     0.7      7000  |  0.7     49000
     0.8      8000  |  0.8     64000
     0.9      9000  |  0.9     81000
     1.0     10000  |  1.0    100000

Note also that this does not affect existing drumkits because each
instrument saves the number of frames with the file.  This only
affects the behavior if you twist a knob.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@916 3b3fb362-3133-0410-aa15-cf69e0a59cb7