Add tests and some fixes to SimpleTransportMaster.
* Modified behavior of normalize() to *NOT* adjust
  the frame.  After further thought, this was the
  wrong thing to do.

* Change SimpleTransportMaster::processed_frames() to
  calculate more stabily... but this is still some
  scaffolding code to get things going.

* add a TransportPosition::normalize(target_frame)
  method to do a normalization, but also make sure that
  we end up on the specified beat.

Status: Hydrogen will scroll when you press play, but still no audio
output.  Also, the GUI seems to be off-by-one when it comes to bars
and beats.  Fails on a vector bounds check in the GUI thread.

Also, The new tests have some tests that fail even though things are
working as expected.  Need to go in and fix up those tests.

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/transport_redesign_2@1151 3b3fb362-3133-0410-aa15-cf69e0a59cb7