Set up a "short circuit" inside audioEngine_process().
If Hydrogen is unable to process audio (e.g. when loading drum kit
samples), then the engine state should be less than STATE_READY.
However, if the state is STATE_PREPARED, then there is an audio driver
that is calling audioEngine_process()... and we need to respond in
real time... and we need to clear out the audio buffers.

To accomplish this, a mutex is added to protect the audio driver's
pointer.  This is locked in fewer places, and allows us to quickly
clear out audio buffers without locking the Big Lock
(AudioEngine::lock()).

Other, non-obvious changes:

* QString's are a minor problem inside the audio threads, especially
  if their constructors are called.  In a few places, I created a
  static QString to help speed things up where we don't need
  dynamic content.

* Some of the logic looking for STATE_READY -or- STATE_PLAYING
  have been simplified to use >, >=, <, <= logic.

* When starting up, some of the states are getting set a little
  early when you consider that the Audio Driver is going to start
  audioEngine_process() immediately after connect().  So, some
  of the state changing logic has been moved around in the init code.

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/jack_zombies@1195 3b3fb362-3133-0410-aa15-cf69e0a59cb7