Test jack port pointer for null before getting track-out pointer.
When per-track outputs are enabled and the user changed drumkits,
sometimes Hydrogen will crash by asking JACK to supply the buffer for
a port, but the port pointer will be null or invalid.  This patch
tests if the jack_port_t pointer is valid before getting its buffer.
When removing the port, the pointer is nulled immediately before
de-registering it.

This patch also makes sure that the per-track port pointers are always
zero when they should be.

git-svn-id: http://svn.assembla.com/svn/hydrogen/trunk@1281 3b3fb362-3133-0410-aa15-cf69e0a59cb7