Make libTritium a shared library.
Doing this involved refractoring interaction between the Playlist and
HydrogenApp.  Playlist was instancing HydrogenApp... which required
libTritium to be linked to all of the GUI code.  This works OK for a
static library, but a shared library won't be able to resolve those
symbols at run time.

To work around it, I created a callback/signal mechanism
(PlaylistListener) that allows HydrogenApp to anonymously listen to
the status of the Playlist.