Refactor build system for handling dependencies.
Beforehand, the build depended entirely on pkg-config... which fails
on a Windows system with no recovery.  This commit adds several
Find*.cmake modules for every package that doesn't come with CMake.

Note that FLAC/FLAC++ provides a bad flac.pc/flac++.pc file to
pkg-config, which causes a problem when building.[1] This has been
worked around.

[1] http://lists.xiph.org/pipermail/flac-dev/2008-November/002602.html