Refactor TransportPosition::bbt_offset to double (was uint32_t).
When I originally used an integer, I spent countless hours combatting
frame drift when using ++/--.  This included establishing the use of
dither() to mitigate frame drift, and several hours of empherical
tests and mathematical theorizing to find a boundary for the amount of
drift to expect.

Using a floating-point type for bbt_offset is not only able to be more
frame-accurate, but almost eliminates the need for the dither()
concept.

While the tests were updated to prove that there is no regression,
they were /not/ updated to prove the new level of accuracy available
with the floating-point bbt_offset.