Modify copyright notices to follow FSF guidelines.
Two not-so-easy to find things done:

* I added my name to Tritium/src/hydrogen.cpp, since
  I've rewritten a majority of the file.

* I've replaced Comix with my name on files that I
  wrote from scratch.

* Added a copyright for Sebastian Moors in timehelper.h
  and timehelper.cpp

As it happens, the Hydrogen project (me included) didn't follow the
FSF guidelines regarding copyright... or there was an implied
assignment to Comix.  For the most part, I have not tried to correct
this with exception of my own work.

Worth considering for the future: getting all copyrights assigned to
the FSF.