Create a "Worker Thread" and refactor Logger.
This change may not make a lot of sense... so here's the idea:

+ The Serialization code is getting segregated.
+ I want to call Serialization code any time,
  anywhere.  So, I'm setting it up to be
  asynchronous.  (Issue command along with a
  callback for when it's complete.)
+ To make it work asynchronously, I need a
  "WorkerThread" that can always respond to
  the work orders.

Serialization isn't the only thing that could benefit from this.  It
could be a whole non-realtime event loop... like the Logger thread.

So, this commit is a step toward that goal, and does the following:

+ Creates a WorkerThread class.
+ Refactors Logger to use it.
+ Moves SongReader into the Serialization code
  (as a temporary step).
+ Corrects a bug that depended on the time delay
  when the logger thread was shutting down
  (sleep(1)).

Note that this commit still passes all tests.