Add member Tritium::LocalFileMng::m_engine.
This requires that LocalFileMng always be instanced with a valid
pointer to an engine.  To accomplished this, several of the
serializable classes needed 'engine' parameters added to their
load/save methods.

The preferred method would be to refactor the serialization code so
that load/save operations happen independently of the engine.
However, I decided to implement this somewhat ugly solution /first/.
Then, I can write tests based on this behavior.  Afterwords I can
refactor the code with some system to detect regressions.