Start using shared pointers in Tritium::Engine.
Working through the serialization code and issues with object
ownership, it became clear that it was time to implement shared
(smart, reference-counted) pointers and let the objects (mostly)
manage themselves rather than using documentation to express change of
ownership.  All of the major objects in Tritium::Engine are now
managed using a shared pointer.

The shared pointers are abstracted through a class template
Tritium::T<> (defined in Tritium/memory.hpp).  So, instead of writing
this:

    boost::shared_ptr<Foo> obj( new Foo );

it is written like this:

    T<Foo>::shared_ptr obj( new Foo );

This abstraction comes at no cost and allows us to change the
implementation of the shared pointer later without changing much
code.  This also avoid possible problems of using a macro like:

    #define SharedPtr boost::shared_ptr

Which is what is done in Raul/Ingen, which is being considered for
future use with Tritium.

While object persistence is more stable, one issue that comes up is
deleting objects that really need to be deleted.  The only case where
this is /really/ important is with the Audio/Midi drivers... where we
need to make sure the driver is shut down so it will release the
hardware.  It's not clear how big an issue this is or will be, yet.
But, it is possible to replace the shared_ptr with something else that
allows more introspection.

The global Tritium::g_engine was not converted to a shared
pointer... nor any reference to Tritium::Engine itself.  I didn't
think it was necessary... but I may be wrong.  We'll see later.

To test for some object persistence issues, some of the tests now
look a little convoluted in the Fixtures.  What used to be declared as
"Foo x;" is now declared "T<Foo>::shared_ptr x_ptr; Foo& x;" so that
the life of the object can be controlled.  E.g. if there are notes
inside a SeqScript, then the Instrument pointers won't be deleted
until the SeqScript is deleted.

Note:  Since I was in the area, I went ahead and removed some of the
dead code in Engine.cpp.  The code has morphed enough that it's not
certain how helpful it will be to refer to those.

Finally, the use of managed pointers is pretty viral.  This "simple"
change resulted in about 2200 lines of code being touched.