Song::save() was calling `delete this`
The SongWriter::writeSong() method originally required a Song* as a
parameter.  When these were changed to T<Song>::shared_ptr, so was
this.  So, in Song::save() the following was used:

    err = writer.writeSong(engine, T<Song>::shared_ptr(this), fn);

This creates a temporary shared pointer to `this`.  When the temporary
variable goes out of scope it deletes the pointer (`delete this`).
This error would manifest itself in the 'autosave' where Composite
would crash while playing a song.

The simplest way to correct the error is to pass a reference (Song&)
instead of a pointer.