Add class to manage object transfer (ObjectBundle).
The ObjectBundle class is a container for several objects (like Song,
Pattern, Instrument, etc.) for the purpose of passing around related
groups.  For example, the intent for the future is that a song be
loaded and the Song and Instruments be passed to Engine after they are
initialized.  Engine will then "install" them.  This is one of the
features that will enable the serialization to code to load a song and
instruments in a non-audio thread, and pass the data when it's ready
to go.