Refactor Tritium::SongReader with Tritium::Serializer.
Serializer (and its buddies) are a system to support asynchronous file
save/restore.  This is being done because it's an important feature to
Composite, and also because the act of moving the Instrument List into
the Sampler would require a bunch of hacking on this section of code,
anyway.

Objects are passed asynchronously using Tritium::ObjectBundle.  This
is a callback/functor with the payload of Song and Instrument
pointers.  This is also the reason for implementing shared pointers.

A summary of the new API may be seen in SongReader::readSong()
[src/Tritium/src/Serialization.cpp, line 900].