JackOutput accidentally activated the jack client pointer too early.
This created a sort of deadlock.
EnginePrivate::audioEngine_startAudioDrivers() would lock the mutex to
the audio driver.  When the audio driver is initialized, it would
activate the jack client... which uses the
EnginePrivate::audioEngine_process_clearAudioBuffers() -- and which
locks the same mutex.  All well and good... except that...

The mutex takes too long, and the jack server terminates the thread.
When audioEngine_startAudioDrivers() unlocks the thread, it tries to
signal the wait condition to the locking instance in
audioEngine_process_clearAudioBuffers().  But, since the thread was
terminated... this is now an invalid memory location and the program
hangs.