Change the test drumkit.
The test drumkit was a symbolic link to the GMkit.  However, when
saving, there have been plenty of schema changes since that drumkit
was originally created.  So, I created a test_kit that is a "fresh
save" of the GMkit.  All the samples for the kit are still symbolic
links.

Note that this creates new failures on the t_Serialization test
because things like the drumkit name don't match.