Sampler no longer queries Preferences.
Instead, the sampler has some configuration features.  If need be,
Preferences should query the sampler in the future.  The main things
that the Sampler was using from Preferences was the maximum notes
(polyphony) setting, whether or not to have per-instrument outputs,
and whether those outputs should be pre- or post- fader.

Also did some code reformatting in Sampler.cpp.