Mixer is now mixing for the Sampler.
The Sampler was taking out a single 'main out' port from the Sampler.
It now takes out a port for each instrument... and the mixer handles
mixing of the instruments.

Note that to test this commit, you'll need to patch
MixerImplPrivate.hpp so that the default send gain is 1.0 (instead of
0.0).