Commit LV2 Plugin of Sampler.
This plugin has most of the internals, but at the moment only outputs
white noise (as a debugging tool).  There is a simple in-tree Makefile
that probably won't work for everyone.  This sampler is currently
unable to load misc. drumkits, nor is it able to accept any input data
(MIDI).  However, it is able to successfully build, install, load
(ingen, zynjacku), and output noise.