Set LINK_INTERFACE_LIBRARIES on libTritium.
CMake cascades link libraries.  So, the sampler and the gui were
getting linked with every library that libTritium was getting linked
to.  This is totally wrong for the LV2 plugin, which should /not/ be
linked to FLAC, sndfile, JACK, etc.  This CMake property overrides
this behavior for both the GUI and the LV2 plugin.