Rearrange the tritium-1 XML Schema.
Documents were only properly validating if the document-level element
was <tritium>.  If it was <programs>, then the schema was allowing
anything.  The reason is that XSD's approach to modular stylesheet
definition is to reuse 'types', whereas I was used to Relax NG where
you could actually re-use element definitions (which is what I really
wanted).