Fix bug in default presets check.
QFileInfo caches its information based on when the object is created.
After creating the object, we actually created the file.  In order for
f_info.exists() to return correctly, we needed to refresh the object.

This bug causes composite_sampler to /not/ load any drumkits on the
very first run.  It also prints a warning message that
tritium:default/presets-sampler could not be loaded.  On the second
run, the issue cleared up.