Add note in INSTALL.txt that in-tree builds are not supported.
An in-tree build of Composite (i.e. 'cmake .') is not supported.  It
may be some time in the future, but not today.  I think the reason
that it fails is because of the filename globbing that is done.  When
cmake re-runs, it finds some cmake-generated source files (like
version.cpp), and this causes problems.

In the future, it's planned to do "make lists" like CMake actually
intended.

Thanks to Harry Van Haaren for uncovering this!