Change libarchive to default (instead of libtar, libz).
libtar is an abandoned library.  Further, its design is not
appropriate for use within a shared library (as I've read... because
of liberal use of global variables).  On many Linux distros that have
not patched libtar, this creates a build error.

Thanks to Alex Stone <compose59@gmail.com> for reporting the issue.