t_AudioPort: fix degenerate tests on float precision.
The original t_AudioPort expected a floating point eps of < 1.0e-45.
This is actually unreasonable for a 32-bit float, which usually has
24-bits of precision.  I.e. when close to 1.0, the minimum precision
is 1.0/(2^24) = 5.96e-08.  So, this test is updated to use twice that:
12.0e-08.

In addition, may of the tests were thing like:

   right[k] = (1.0f/left[k])

However, it's more complex to prove the precision of this calculation.
So, these have been changed to a simpler:

   right[k] = (1e-6f * left[k])

Which should always be within the 12.0e-08 precision.