Delete all the CPack stuff.
I never used it, and I tried once to use it on Windows.  I hated it.
You have to do so much platform-specific stuff and overrides, you
might as well package it for each platform.  In addition, the installer
for CPack on windows in Nullsoft -- which I was very unhappy with.
I prefer an MSI installer (like WiX) because it handles uninstalls
much better.