Build system: Convert from GLOB to actual file lists.
Source files had been found by a GLOB_RECURSE function.  The up-side
is that the developer does not have to manually add files to the
CMakeLists.txt file.  The down-side is that he has to manually re-run
CMake in order to detect the new sources.

The Tritium tests still use GLOB_RECURSE for headers, but that's an
exception.