Rearrange the Composite source directory.
Composite's directory tree /was/ arranged like Tritiums, where you
have:

   Composite/
   +-- Composite/
   +-- src/

Because I was dividing source files into folders with the same
namespace, this became:

   Composite/
   +-- Composite/
   |   +-- Main/
   |   +-- Widgets/
   |   +-- etc...
   |
   +-- src/
       +-- Main/
       +-- Widgets/
       +-- etc...

This parallelism was getting annoying when adding a new class.
Therefore, the directories are combined.  If we ever get serious about
exporting Composite headers... then anything marked "Private" will be
excluded.

This currently can't be done for Tritium because there are a lot of
private classes in the source folder that aren't marked 'Private'.