Fix FTBFS: t_AudioPort test Missing link to QtCore
On Ubuntu/Natty, the Tritium tests would fail because of a missing
link to QtCore.  This happened because it pulled in a private header
from the Tritium source tree.  Note that apps that link to libTritium
should /not/ have to directly link to Qt.

Patch courtesy of Alessio Treglia <alessio@debian.org> 2011-02-28