LV2 Midi Event: Check event type for MIDI events
Beforehand, the Composite sampler assumed that all events being sent
to the sampler were MIDI events <http://lv2plug.in/ns/ext/midi#MidiEvent>.
However, this is a false assumption since the documentation says that
the type must be compared with the value returned from the uri-map
feature.

There have been no known crashes due to this, but there is the
potential for one.