Remove all references to QCoreApplication.
To facilitate development, we were using QCoreApplication to find a
folder called "data/" in the same folder as the executable.  This is a
leftover from Hydrogen and is no longer even valid.  In addition,
using QCoreApplication in Tritium (and the LV2 Sampler) is buggy.

Removing QCoreApplication from Tritium.  As a consequence, you must
have Tritium/Composite installed (somewhere) in order to test it or it
will not find its resources.