Tritium: Effects::getPluginList() was constantly scanning directories.
When the user has no plugins installed, Effects::getPluginList() will
constantly search folders for them.  This is a problem because it's
being called from MixerImpl::mix_down() (which expects that it is
receiving cached data).

Adding a class variable to indicate that the info has been cached.
Also returning a reference instead of a copy.