composite.desktop.in: Fixing incorrect 'Version' value
According to the Desktop Entry Specification[1] the 'Version' key
refers to the version of the Desktop Entry Specification that the file
is conforming to.[2] It is /not/ the version of the application.
However, this was being used as the version of Composite.

Setting 'Version' to 1.0 to conform with the spec.

Thanks to Dominique Michel <dominique.michel@vtxnet.ch> for reporting
this!

[1] http://standards.freedesktop.org/desktop-entry-spec/latest/
[2] http://standards.freedesktop.org/desktop-entry-spec/latest/ar01s05.html

Signed-off-by: Gabriel M. Beddingfield <gabrbedd@gmail.com>